+++ src/main/java/uk/gov/hmcts/darts/audio/service/MediaRequestService.java
@@ -3,7 +3,7 @@ package uk.gov.hmcts.darts.audio.service;
 
 import uk.gov.hmcts.darts.audio.entity.MediaRequestEntity;
 import uk.gov.hmcts.darts.audio.enums.AudioRequestStatus;
-import uk.gov.hmcts.darts.audiorequest.model.AudioRequestDetails;
+import uk.gov.hmcts.darts.audio.model.AudioRequestDetails;
 
 public interface MediaRequestService {
