+++ src/main/resources/openapi/audio.yaml
